[Refactor #6] Split CPU module test code
