+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod cpu;
+pub mod opcodes;
+
+#[macro_use]
+extern crate lazy_static;
